Fix repo mapping metadata section:
Mapping metadata is required on all full verification ECUs, not just
Primaries.  While FV secondaries might not require the URLs for each
repository, they do need to know which repositories to expect
matching metadata from in order to trust a given image.  That info
comes from the mapping metadata.

Also explain those two purposes of the mapping metadata (NOTE: This
likely aids understanding, but might be inappropriate in this
document!)

Clarify the verificaiton requirement in terms of the client, which
is more what is really happening.  (We're not enforcing a policy
on the repository; we're enforcing a verification policy on the
client.)

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>